BE/pyproject.toml
[project]
name = "chores"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.121.1",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "ruff>=0.14.4",
    "sqlmodel>=0.0.27",
]
